+++ source/class/qx/tool/cli/commands/package/Migrate.js
@@ -15,54 +15,19 @@
      * Christian Boulanger (info@bibliograph.org, @cboulanger)
 
 ************************************************************************ */
-const process = require("process");
 
 /**
- * Installs a package
+ * @deprecated {7.0.0}
  */
 qx.Class.define("qx.tool.cli.commands.package.Migrate", {
-  extend: qx.tool.cli.commands.Package,
-
+  extend: qx.tool.cli.commands.Migrate,
   statics: {
-    /**
-     * Flag to prevent recursive call to process()
-     */
-    migrationInProcess: false,
-    /**
-     * Return the Yargs configuration object
-     * @return {{}}
-     */
-    getYargsCommand: function() {
-      return {
-        command: "migrate",
-        describe: "migrates the package system to a newer version.",
-        builder: {
-          "verbose": {
-            alias: "v",
-            describe: "Verbose logging"
-          },
-          "dry-run": {
-            describe: "Do not apply migrations"
-          }
-        }
-      };
-    }
+    getYargsCommand: qx.tool.cli.commands.Migrate.getYargsCommand
   },
-
   members: {
-    /**
-     * Announces or applies a migration
-     */
     process: async function() {
-      let runner = new qx.tool.migration.Runner().set({
-        command: this,
-        dryRun: this.argv.dryRun,
-        verbose: this.argv.verbose
-      });
-      runner.runMigrations();
-      if (!this.argv.quiet) {
-        qx.tool.compiler.Console.info("Migration completed.");
-      }
+      qx.tool.compiler.Console.warn("`*** qx package migrate` has been deprecated in favor of `qx migrate`.");
+      this.base(arguments);
     }
   }
 });