+++ CHANGELOG.md
@@ -10,8 +10,11 @@ source, add the path to `compile.json`'s `libraries` array.
 The maximum value is at the top of the slider and the minimal value
 at the bottom. Before that, the 2 values were reversed 
 
-- all flash
-supporting classes are removed - flash is dead since January 2020.
+- all flash supporting classes are removed - flash is dead since January 2020.
+
+## Deprecations:
+
+- `qx package migrate` has been deprecated in favor of `qx migrate` 
 
 # v6.0.0